cyclic import: files/modules/classes re arranged
