+++ tuna/miopen/subcmd/export_db.py
@@ -39,7 +39,8 @@ from tuna.miopen.db.miopen_tables import GoldenMixin
 from tuna.miopen.db.tables import MIOpenDBTables
 from tuna.miopen.utils.metadata import SQLITE_PERF_DB_COLS
 from tuna.utils.db_utility import DB_Type
-from tuna.miopen.utils.miopen_db_utils import get_id_solvers
+#from tuna.miopen.utils.miopen_db_utils import get_id_solvers
+from tuna.miopen.db.solver import get_id_solvers
 from tuna.utils.logger import setup_logger
 from tuna.miopen.utils.analyze_parse_db import get_config_sqlite, insert_solver_sqlite
 from tuna.miopen.utils.analyze_parse_db import get_sqlite_cfg_dict
