+++ tuna/miopen/subcmd/load_job.py
@@ -36,7 +36,7 @@ from sqlalchemy.exc import IntegrityError  #pylint: disable=wrong-import-order
 from sqlalchemy.sql.expression import true
 
 from tuna.miopen.utils.metadata import ALG_SLV_MAP, TENSOR_PRECISION
-from tuna.miopen.utils.miopen_db_utils import get_solver_ids
+from tuna.miopen.db.solver import get_solver_ids
 from tuna.utils.logger import setup_logger
 from tuna.utils.db_utility import connect_db
 from tuna.miopen.db.miopen_tables import Solver
@@ -44,6 +44,8 @@ from tuna.dbBase.sql_alchemy import DbSession
 from tuna.miopen.utils.config_type import ConfigType
 from tuna.miopen.db.tables import MIOpenDBTables
 from tuna.miopen.parse_miopen_args import get_load_job_parser
+
+#pylint: disable=R0914
 
 
 def arg_fin_steps(args: argparse.Namespace):
@@ -154,7 +156,9 @@ def add_jobs(args: argparse.Namespace, dbt: MIOpenDBTables,
     fin_step_str = 'not_fin'
     if args.fin_steps:
       fin_step_str = ','.join(args.fin_steps)
-    query = f"select config, solver from {dbt.job_table.__tablename__} where session={args.session_id} and fin_step='{fin_step_str}'"
+    query = f"select config, solver from {dbt.job_table.__tablename__} \
+      where session={args.session_id} and fin_step='{fin_step_str}'"
+
     logger.info(query)
     ret = session.execute(query)
     pre_ex: Dict[str, Dict[str, bool]] = {}
