+++ tuna/miopen/worker/fin_class.py
@@ -50,7 +50,8 @@ from tuna.miopen.worker.fin_utils import get_fin_slv_status
 from tuna.miopen.utils.config_type import ConfigType
 from tuna.miopen.utils.parsing import parse_pdb_key
 from tuna.utils.db_utility import session_retry
-from tuna.miopen.utils.miopen_db_utils import get_solver_ids, get_id_solvers
+#from tuna.miopen.utils.miopen_db_utils import get_solver_ids, get_id_solvers
+from tuna.miopen.db.solver import get_solver_ids, get_id_solvers
 from tuna.utils.db_utility import gen_select_objs, gen_insert_query, gen_update_query
 from tuna.utils.db_utility import get_class_by_tablename, has_attr_set
 from tuna.utils.utility import split_packets
@@ -62,6 +63,7 @@ class FinClass(WorkerInterface):
 
   # pylint: disable=too-many-instance-attributes
   # pylint: disable=too-many-public-methods
+  # pylint: disable=no-member
 
   def __init__(self, **kwargs):
     """Constructor"""
